Button clicking, errors fixing
I tried to work on the add button without any error and/or warning messages after create the minigames menu. I ended up fixing a lot of errors I got after I started to build the program. (in Games- and Library Fragment, I put the error part in a comment and I fixed SmoothGreen to smoothGreen in listview_selector.

(I still cannot test if my button is working and it gives the message after you click, so if someone could test it after he pulled this change, that would be lovely!)